Switch to bluebird, and fix annoying hangs.
This tool was frequently hanging in a mysterious way. I was having a hard time
debugging it, and switched from Q to bluebird, which did help a little in the debugging,
though it wasn't by any means the magic bullet.

Ultimately I think the worst problem is that the logic I had wasn't exposing
pivotal api requests that were failing with error messages, but there was also
some weird interaction with readline.